4-email-hash-field: move argon2hash and token to krypto package
