+++ internal/auth/user.go
@@ -4,13 +4,14 @@ import (
 	"time"
 
 	"github.com/willemschots/househunt/internal/email"
+	"github.com/willemschots/househunt/internal/krypto"
 )
 
 // User contains the data for a user.
 type User struct {
 	ID           int
 	Email        email.Address
-	PasswordHash Argon2Hash
+	PasswordHash krypto.Argon2Hash
 	IsActive     bool
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
